requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
transformers
sentence-transformers
faiss-cpu
python-multipart
uvicorn
pydantic
aiohttp
